feat(api): return typed EthBlock responses and refactor block handling
- change BlockResource to return List<EthBlock> instead of raw JSON strings
- add ObjectMapper-based deserialization from stored block data
- introduce detailed EthBlock DTO with nested structures
- switch block number types from int to long across entities and models
- improve scanner logging with formatted range output
- update docker volume name (mariadb_data → db_data)
- adjust scheduler interval and set custom HTTP port

+++ src/main/java/com/simorghsoftech/core/models/Block.java
@@ -4,17 +4,13 @@ package com.simorghsoftech.core.models;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.web3j.utils.Numeric;
-
-import java.math.BigInteger;
-import java.util.Objects;
 
 public class Block {
-    private final int number;
+    private final long number;
     private final String hash;
     private final String data;
 
-    public Block(int number, String hash, String data) {
+    public Block(long number, String hash, String data) {
         this.number = number;
         this.hash = hash;
         this.data = data;
@@ -28,7 +24,7 @@ public class Block {
         return hash;
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
